refactor: change id variable type to long

+++ src/main/java/com/omelentjeff/citybike/controller/JourneyController.java
@@ -32,7 +32,7 @@ public class JourneyController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<JourneyDTO> findJourneyById(@PathVariable int id) {
+    public ResponseEntity<JourneyDTO> findJourneyById(@PathVariable long id) {
         return ResponseEntity.ok(journeyService.getJourneyById(id));
     }
 }
